Added update functionality if newer version of feed is available
